additional timestamp refactor
