+++ libs/model/src/lib/timestamp.ts
@@ -1,3 +1,25 @@
-import type firestore from 'firebase/firestore';
+import type { Timestamp } from 'firebase/firestore';
 
-export type Timestamp = firestore.Timestamp;  // TODO #8250
+export { Timestamp };
+
+
+/** Checks if the date is a firestore Timestamp. */
+export function isTimeStamp(date: Timestamp | Date): date is Timestamp {
+  return date && !(date instanceof Date);
+}
+
+/** Takes a Date, a string or a Timestamp and returns it as a Date. */
+
+export function toDate(date: Timestamp | Date): Date {
+  if (isTimeStamp(date)) {
+    return date.toDate();
+  }
+  if (typeof date === 'string' || typeof date === 'number') {
+    return new Date(date);
+  }
+  if (!date) {
+    // Return null to avoid undefined error with firestore
+    return null;
+  }
+  return date;
+}
