+++ libs/utils/src/lib/helpers.ts
@@ -38,28 +38,6 @@ export function mergeDeep<T>(target: T, source: Partial<T>): T {
   return output;
 }
 
-/** Checks if the date is a firestore Timestamp. */
-function isTimeStamp(
-  date: Timestamp | Date
-): date is Timestamp {
-  return date && date instanceof Timestamp; // TODO #8250
-}
-
-/** Takes a Date, a string or a Timestamp and returns it as a Date. */
-export function toDate(date: Timestamp | Date): Date {
-  if (isTimeStamp(date)) {
-    return date.toDate();
-  }
-  if (typeof date === 'string' || typeof date === 'number') {
-    return new Date(date);
-  }
-  if (!date) {
-    // Return null to avoid undefined error with firestore
-    return null;
-  }
-  return date;
-}
-
 /**
  * Get the value of an item based on a path
  * @example item = movie, key = 'budget.totalBudget'
